+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Next
 - Modernize the code base, testing with Python 3.10 and 3.11
 - Add UDP server with support for custom protocols
 - Add protocol for the Micron Bolt Mini 2 GPS tracker
+- New filter: YAML/JSON parser
 
 Version 0.1.16 - 2021-06-23
 ===========================